22º - TESTS.Produtos - Testeo general en productos
Se han añadido test a los productos los cuales comprueban el corrrecto funcionamiento del crud al completo

También se ha controlado el acceso a la creación de productos, tanto para los Guests como para los usuarios autenticados.

De igual manera, se ha controlado la edición y eliminación de productos para que sola y exclusivamente los usuarios autores del producto puedan editar y eliminar el mismo (evidentemente, los Guest y los usuarios no propietarios no podrán hacer cambios, esto conlleva que ni si  quiera puedan acceder al formulario o ver los botones correspondientes de edicion o eliminación, ademas de que quedan restringidos de usar peticiones Patch, Post o Delete)

+++ app/Http/Controllers/ProductController.php
@@ -39,10 +39,17 @@ class ProductController extends Controller
         return view('product.product-show', compact('product'));
     }
     public function edit(Product $product) {
+        if ($product->seller_id != Auth::id()){
+            return redirect()->route('product.show', $product)->with('status', 'No puedes editar un articulo que no es tuyo');
+        }
         $categories = Category::all();
         return view('product.product-edit', compact('product', 'categories'));
     }
     public function update(ProductUpdateRequest $request, Product $product) {
+
+        if ($product->seller_id != Auth::id()){
+            return redirect()->route('product.show', $product)->with('status', 'No puedes editar un articulo que no es tuyo');
+        }
 
         $validated = $request->validated();
 
@@ -59,11 +66,15 @@ class ProductController extends Controller
 
 
         $product->update($validated);
-        return redirect()->route('product.show', $product)->with('status', 'Product updated successfully');
+        return redirect()->route('product.show', $product)->with('status', 'Producto actualizado correctamente');
     }
     public function destroy(Product $product) {
+        if ($product->seller_id != Auth::id()){
+            return redirect()->route('product.show', $product)->with('status', 'No puedes eliminar un articulo que no es tuyo');
+        }
+
         ProductImageService::destroyAllImages($product);
         $product->delete();
-        return redirect()->route('product.index')->with('status', 'Product deleted successfully');
+        return redirect()->route('product.index')->with('status', 'El producto ha sido eliminado correctamente');
     }
 }
